VFO colors updated

+++ src/display/ILI9488/Display-ILI9488-320x480.hpp
@@ -31,7 +31,7 @@ private:
   void refreshVFOMode(uint8_t number, bool isActive, TRXVFOMode mode);
   void refreshVFOStep(uint8_t number, bool isActive, uint64_t step);
   void createDigitalSMeter();
-  void refreshRNDDigitalSMeter(int newSignal);
+  void refreshRNDDigitalSMeter(uint8_t newSignal);
   void refreshVolume(uint8_t AFGain, bool isMuted);
   void refreshPassBandFilter(uint32_t LF, uint32_t HF, uint32_t BW);
 };
